+++ CHANGELOG.md
@@ -4,7 +4,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## Unreleased
+## [1.0.0-beta.2] - Unreleased
+### Added
+- Add `astarte_device_stop` function to stop the internal MQTT client.
+- Add connection and disconnection callbacks called after device connection/disconnection.
+
+### Changed
+- `astarte_device_start` now returns `astarte_err_t` to check if the device was succesfully started.
+
+## [1.0.0-beta.1] - 2021-02-16
 ### Added
 - Allow sending data to object aggregated interfaces.
 - Allow using UUIDv5 to derive the hardware id.